+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Significantly improved video slider seek experience
 * Fixed video pause/play to trigger on click down instead of button release
 * Added F5 keyboard shortcut for refresh
+* Added scroll up down buttons
 
 # v4.29.1